pyproject.toml
[tool.poetry]
name = "ripix"
version = "2.2.0"
description = "A Rich-compatible library for writing pixel images and ASCII art to the terminal."
authors = ["Darren Burns <darrenb900@gmail.com>", "Romanin <semina054@gmail.com>"]
readme = "README.md"
packages = [{ include = "ripix" }]
include = ["ripix/py.typed", "ripix/functions.pyi"]

[tool.poetry.dependencies]
python = "^3.7"
rich = ">=12.0.0"
pillow = "^9.0.0"
asyncio = "^3.4.3"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
mypy = "^0.990"
syrupy = "^3.0.5"
importlib-metadata = "^5.0.0"
types-pillow = "^9.3.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
